rest of screens reworked to MVVM
